chore(bench): separate competitor benchmarks from CodSpeed and switch to walltime
Split benchmarks into focused files for clearer regression tracking:
- distance.bench.ts: edit distance (levenshtein, damerau, hamming)
- similarity.bench.ts: similarity scores (jaro, jaroWinkler, sorensenDice)
- ratio.bench.ts: token-based ratios (tokenSetRatio, tokenSortRatio, weightedRatio)
- batch.bench.ts: 1:N operations (levenshteinMany, normalizedLevenshteinMany)
- search.bench.ts: fuzzy search, closest match, index construction

Move competitor benchmarks to *.compare.bench.ts files (local use only,
excluded from CodSpeed via bench:ci script).

Switch CodSpeed from simulation to walltime mode for more realistic
performance measurements.

Closes #371
