+++ __test__/search.bench.ts
@@ -1,9 +1,3 @@
-import uFuzzy from '@leeoniya/ufuzzy';
-import { closest as fastestLevenshteinClosest } from 'fastest-levenshtein';
-import { Index as FlexSearchIndex } from 'flexsearch';
-import Fuse from 'fuse.js';
-import fuzzysort from 'fuzzysort';
-import MiniSearch from 'minisearch';
 import { bench, describe } from 'vitest';
 
 import { closest, FuzzyIndex, search } from '../index.js';
@@ -139,46 +133,6 @@ const hugeItems = Array.from({ length: 100_000 }, (_, i) => {
 
 // --- Pre-initialize search instances ---
 
-// Fuse.js
-const fuseSmall = new Fuse(smallItems, { threshold: 0.4 });
-const fuseMedium = new Fuse(mediumItems, { threshold: 0.4 });
-const fuseLarge = new Fuse(largeItems, { threshold: 0.4 });
-
-// fuzzysort (prepared targets)
-const fuzzysortMediumPrepared = mediumItems.map((item) => fuzzysort.prepare(item));
-const fuzzysortLargePrepared = largeItems.map((item) => fuzzysort.prepare(item));
-const fuzzysortXlargePrepared = xlargeItems.map((item) => fuzzysort.prepare(item));
-const fuzzysortHugePrepared = hugeItems.map((item) => fuzzysort.prepare(item));
-
-// uFuzzy
-const uf = new uFuzzy();
-
-// FlexSearch (default strict tokenization — represents inverted-index search category,
-// intentionally not configured for fuzzy/typo-tolerant matching)
-const flexSmall = new FlexSearchIndex();
-for (let i = 0; i < smallItems.length; i++) flexSmall.add(i, smallItems[i]);
-const flexMedium = new FlexSearchIndex();
-for (let i = 0; i < mediumItems.length; i++) flexMedium.add(i, mediumItems[i]);
-const flexLarge = new FlexSearchIndex();
-for (let i = 0; i < largeItems.length; i++) flexLarge.add(i, largeItems[i]);
-const flexXlarge = new FlexSearchIndex();
-for (let i = 0; i < xlargeItems.length; i++) flexXlarge.add(i, xlargeItems[i]);
-const flexHuge = new FlexSearchIndex();
-for (let i = 0; i < hugeItems.length; i++) flexHuge.add(i, hugeItems[i]);
-
-// MiniSearch
-function createMiniSearch(items: string[]) {
-  const ms = new MiniSearch({ fields: ['text'], storeFields: ['text'] });
-  ms.addAll(items.map((text, id) => ({ id, text })));
-  return ms;
-}
-const miniSmall = createMiniSearch(smallItems);
-const miniMedium = createMiniSearch(mediumItems);
-const miniLarge = createMiniSearch(largeItems);
-const miniXlarge = createMiniSearch(xlargeItems);
-const miniHuge = createMiniSearch(hugeItems);
-
-// rapid-fuzzy FuzzyIndex
 const fuzzyIndexSmall = new FuzzyIndex(smallItems);
 const fuzzyIndexMedium = new FuzzyIndex(mediumItems);
 const fuzzyIndexLarge = new FuzzyIndex(largeItems);
@@ -195,26 +149,6 @@ describe('Fuzzy Search — Small (20 items)', () => {
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     fuzzyIndexSmall.search('aple', { maxResults: 5 });
   });
-
-  bench('fuse.js', () => {
-    fuseSmall.search('aple', { limit: 5 });
-  });
-
-  bench('fuzzysort', () => {
-    fuzzysort.go('aple', smallItems, { limit: 5 });
-  });
-
-  bench('uFuzzy', () => {
-    uf.search(smallItems, 'aple');
-  });
-
-  bench('FlexSearch', () => {
-    flexSmall.search('aple', { limit: 5 });
-  });
-
-  bench('MiniSearch', () => {
-    miniSmall.search('aple', { fuzzy: 0.2, prefix: true });
-  });
 });
 
 describe('Fuzzy Search — Medium (1K items)', () => {
@@ -224,26 +158,6 @@ describe('Fuzzy Search — Medium (1K items)', () => {
 
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     fuzzyIndexMedium.search('utils config', { maxResults: 10 });
-  });
-
-  bench('fuse.js', () => {
-    fuseMedium.search('utils config', { limit: 10 });
-  });
-
-  bench('fuzzysort', () => {
-    fuzzysort.go('utils config', fuzzysortMediumPrepared, { limit: 10 });
-  });
-
-  bench('uFuzzy', () => {
-    uf.search(mediumItems, 'utils config');
-  });
-
-  bench('FlexSearch', () => {
-    flexMedium.search('utils config', { limit: 10 });
-  });
-
-  bench('MiniSearch', () => {
-    miniMedium.search('utils config', { fuzzy: 0.2, prefix: true });
   });
 });
 
@@ -255,47 +169,11 @@ describe('Fuzzy Search — Large (10K items)', () => {
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     fuzzyIndexLarge.search('handler middleware', { maxResults: 10 });
   });
-
-  bench('fuse.js', () => {
-    fuseLarge.search('handler middleware', { limit: 10 });
-  });
-
-  bench('fuzzysort', () => {
-    fuzzysort.go('handler middleware', fuzzysortLargePrepared, { limit: 10 });
-  });
-
-  bench('uFuzzy', () => {
-    uf.search(largeItems, 'handler middleware');
-  });
-
-  bench('FlexSearch', () => {
-    flexLarge.search('handler middleware', { limit: 10 });
-  });
-
-  bench('MiniSearch', () => {
-    miniLarge.search('handler middleware', { fuzzy: 0.2, prefix: true });
-  });
 });
 
 describe('Fuzzy Search — Extra Large (50K items)', () => {
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     fuzzyIndexXlarge.search('handler middleware', { maxResults: 10 });
-  });
-
-  bench('fuzzysort', () => {
-    fuzzysort.go('handler middleware', fuzzysortXlargePrepared, { limit: 10 });
-  });
-
-  bench('uFuzzy', () => {
-    uf.search(xlargeItems, 'handler middleware');
-  });
-
-  bench('FlexSearch', () => {
-    flexXlarge.search('handler middleware', { limit: 10 });
-  });
-
-  bench('MiniSearch', () => {
-    miniXlarge.search('handler middleware', { fuzzy: 0.2, prefix: true });
   });
 });
 
@@ -303,67 +181,17 @@ describe('Fuzzy Search — Huge (100K items)', () => {
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     fuzzyIndexHuge.search('handler middleware', { maxResults: 10 });
   });
-
-  bench('fuzzysort', () => {
-    fuzzysort.go('handler middleware', fuzzysortHugePrepared, { limit: 10 });
-  });
-
-  bench('uFuzzy', () => {
-    uf.search(hugeItems, 'handler middleware');
-  });
-
-  bench('FlexSearch', () => {
-    flexHuge.search('handler middleware', { limit: 10 });
-  });
-
-  bench('MiniSearch', () => {
-    miniHuge.search('handler middleware', { fuzzy: 0.2, prefix: true });
-  });
 });
 
 describe('Index Construction — Medium (1K items)', () => {
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     new FuzzyIndex(mediumItems);
   });
-
-  bench('fuse.js', () => {
-    new Fuse(mediumItems, { threshold: 0.4 });
-  });
-
-  bench('fuzzysort (prepare)', () => {
-    mediumItems.map((item) => fuzzysort.prepare(item));
-  });
-
-  bench('FlexSearch', () => {
-    const idx = new FlexSearchIndex();
-    for (let i = 0; i < mediumItems.length; i++) idx.add(i, mediumItems[i]);
-  });
-
-  bench('MiniSearch', () => {
-    createMiniSearch(mediumItems);
-  });
 });
 
 describe('Index Construction — Large (10K items)', () => {
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     new FuzzyIndex(largeItems);
-  });
-
-  bench('fuse.js', () => {
-    new Fuse(largeItems, { threshold: 0.4 });
-  });
-
-  bench('fuzzysort (prepare)', () => {
-    largeItems.map((item) => fuzzysort.prepare(item));
-  });
-
-  bench('FlexSearch', () => {
-    const idx = new FlexSearchIndex();
-    for (let i = 0; i < largeItems.length; i++) idx.add(i, largeItems[i]);
-  });
-
-  bench('MiniSearch', () => {
-    createMiniSearch(largeItems);
   });
 });
 
@@ -375,10 +203,6 @@ describe('Closest Match — Medium (1K items)', () => {
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     fuzzyIndexClosestMedium.closest('src/utils42.ts');
   });
-
-  bench('fastest-levenshtein', () => {
-    fastestLevenshteinClosest('src/utils42.ts', mediumItems);
-  });
 });
 
 describe('Closest Match — Large (10K items)', () => {
@@ -388,9 +212,5 @@ describe('Closest Match — Large (10K items)', () => {
 
   bench('rapid-fuzzy (FuzzyIndex)', () => {
     fuzzyIndexClosestLarge.closest('handler_middleware_500');
-  });
-
-  bench('fastest-levenshtein', () => {
-    fastestLevenshteinClosest('handler_middleware_500', largeItems);
   });
 });
